backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
pymupdf  # PDF text extraction
sentence-transformers  # Semantic search embeddings
faiss-cpu  # Vector search index
openai  # LLM integration
tqdm  # Progress tracking
python-dotenv  # Environment variable management
python-multipart # UploadFile, File, Form in FastAPI
